Replace the string based handling in crunch by a dict based handler
The resulting code actually means crunching takes longer but the benefit
is that it is easier to understand and maintain/update.

Also setting up a new helpers.py to phase out the current Common.py and
containing methods needed for dict based handling.

Results (/data/) are (with exceptions below) equivalent to previous up to:
* order of entries
* order in individual lists
* whitespace which should have been trimmed before but wasn't

The exceptions:
* in photo: same_object sometimes contains no entries. However these photos also
	    contain no objects (also before redux) so likely this was an issue
	    with the older code
* in kuenstler: The redux version fixes a bug which some times caused years to
		be parsed incorreclty.

For comparing the results the following script was used
https://gist.github.com/lokal-profil/ad8c45432b2754278001

closes #3